youtube play button bug fixed with xpath
selenium is not registering click() for certain events when it comes to mobile, so i have fixed them with Xpaths by removing classes. 
removed all time.sleep function calls
initially the script was enabling stats for nerds first and then playing the video, which is not the case with mobile phones. you first play the video and then enable stats for nerds. i have fixed that as well.